add docker setup for db_v2 it test. adjust test calls
